pipeline/training/requirements.txt
google-cloud-aiplatform==1.6.2
google-cloud-storage==1.42.3
kfp==1.8.7
google-cloud-pipeline-components==0.1.9
